+++ packages/checkbox/src/shared/index.ts
@@ -1,0 +1,1 @@
+export { getCheckboxTestIds } from '../utils';