+++ packages/shared/src/index.ts
@@ -13,3 +13,5 @@ export * from './context/PortalContext';
 export * from './warning';
 export * from './object';
 export * from './type-checks';
+export * from './get-color-var';
+export * from './mask';