+++ packages/themes/src/corp.ts
@@ -1,4 +1,0 @@
-// eslint-disable-next-line import/no-webpack-loader-syntax
-import styles from '!!css-loader!@alfalab/core-components/themes/corp.css';
-
-export default styles.toString();