+++ packages/universal-modal/src/components/content/desktop/Component.desktop.tsx
@@ -1,7 +1,10 @@
-import React, { FC } from 'react';
+import React, { FC, useContext } from 'react';
 import cn from 'classnames';
 
+import { UniversalModalContext } from '../../../context/universal-modal-context';
 import { BaseContent, ContentProps } from '../base-content/base-content';
+
+import styles from './desktop.module.css';
 
 export type ContentDesktopProps = ContentProps & {
     /**
@@ -10,6 +13,16 @@ export type ContentDesktopProps = ContentProps & {
     size?: 's' | 500;
 };
 
-export const ContentDesktop: FC<ContentDesktopProps> = ({ className, ...restProps }) => (
-    <BaseContent className={cn(className)} {...restProps} />
-);
+export const ContentDesktop: FC<ContentDesktopProps> = ({ className, ...restProps }) => {
+    const { hasHeader, hasFooter } = useContext(UniversalModalContext);
+
+    return (
+        <BaseContent
+            className={cn(styles.content, className, {
+                [styles.withHeader]: hasHeader,
+                [styles.withFooter]: hasFooter,
+            })}
+            {...restProps}
+        />
+    );
+};