+++ packages/universal-modal/src/components/footer/responsive/Component.responsive.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react';
+import React, { forwardRef } from 'react';
 
 import { useIsDesktop } from '@alfalab/core-components-mq';
 
@@ -7,12 +7,14 @@ import { FooterMobile, FooterMobileProps } from '../mobile/Component.mobile';
 
 export type FooterResponsiveProps = FooterDesktopProps | FooterMobileProps;
 
-export const FooterResponsive: FC<FooterResponsiveProps> = ({ ...restProps }) => {
+export const FooterResponsive = forwardRef<HTMLDivElement, FooterResponsiveProps>((props, ref) => {
     const isDesktop = useIsDesktop();
 
     if (isDesktop) {
-        return <FooterDesktop {...restProps} />;
+        return <FooterDesktop ref={ref} {...props} />;
     }
 
-    return <FooterMobile {...restProps} />;
-};
+    return <FooterMobile {...props} />;
+});
+
+FooterResponsive.displayName = 'FooterResponsive';