+++ packages/universal-modal/src/desktop/utils/isHeaderNode.ts
@@ -1,0 +1,2 @@
+export const isHeaderNode = (displayName: string) =>
+    ['HeaderDesktop', 'HeaderResponsive'].includes(displayName);