+++ packages/universal-modal/src/docs/Component.stories.tsx
@@ -195,7 +195,6 @@ export const mobile: Story = {
             ['start', 'center', 'space-between', 'column'],
             'start',
         );
-        const bigTitle = boolean('bigTitle', false);
         const open = boolean('open', false);
         const appearance = select('appearance', ['bottom', 'right'], 'bottom');
 
@@ -210,26 +209,25 @@ export const mobile: Story = {
                     onClose={() => setOpen(false)}
                 >
                     {header && (
-                        <UniversalModalDesktop.Header
+                        <UniversalModalMobile.Header
                             sticky={stickyHeader}
                             title={headerTitle}
                             align={headerAlign}
-                            bigTitle={bigTitle}
                         />
                     )}
-                    <UniversalModalDesktop.Content>
+                    <UniversalModalMobile.Content>
                         <Content />
-                    </UniversalModalDesktop.Content>
+                    </UniversalModalMobile.Content>
 
                     {stickyFooter && (
-                        <UniversalModalDesktop.Footer sticky={stickyFooter} layout={footerLayout}>
+                        <UniversalModalMobile.Footer sticky={stickyFooter} layout={footerLayout}>
                             <ButtonMobile size={56} view='primary'>
                                 Сохранить
                             </ButtonMobile>
                             <ButtonMobile size={56} view='secondary'>
                                 Отмена
                             </ButtonMobile>
-                        </UniversalModalDesktop.Footer>
+                        </UniversalModalMobile.Footer>
                     )}
                 </UniversalModalMobile>
             </div>