+++ packages/universal-modal/src/mobile/types/props.ts
@@ -11,13 +11,15 @@ export type BaseUniversalModalMobileProps = {
 };
 
 export type UniversalModalMobileProps = BaseUniversalModalMobileProps &
-    Pick<
-        BaseModalProps,
-        | 'children'
-        | 'dataTestId'
-        | 'open'
-        | 'className'
-        | 'wrapperClassName'
-        | 'onUnmount'
-        | 'transitionProps'
+    Pick<BaseModalProps, 'open'> &
+    Partial<
+        Pick<
+            BaseModalProps,
+            | 'children'
+            | 'dataTestId'
+            | 'className'
+            | 'wrapperClassName'
+            | 'onUnmount'
+            | 'transitionProps'
+        >
     >;