+++ packages/base-modal/CHANGELOG.md
@@ -1,4 +1,68 @@
-# Change Log
+# @alfalab/core-components-base-modal
+
+## 6.0.0
+
+### Major Changes
+
+<sup><time>05.08.2025</time></sup>
+
+#### [#1611](https://github.com/core-ds/core-components/pull/1611)
+
+Обновлена сборка.
+
+Добавлены пропущенные зависимости.
+
+Синхронизированы версии зависимостей.
+
+<sup><time>05.08.2025</time></sup>
+
+#### [#1611](https://github.com/core-ds/core-components/pull/1611)
+
+-   Пакет `@alfalab/stack-context` перенесен в `@alfalab/core-components-stack-context`. При обновлении необходимо заменить `as-is`.
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-backdrop@4.0.0
+    -   @alfalab/core-components-global-store@3.0.0
+    -   @alfalab/core-components-portal@4.0.0
+    -   @alfalab/core-components-shared@1.0.0
+    -   @alfalab/core-components-stack@6.0.0
+    -   @alfalab/core-components-stack-context@1.0.0
+
+## 5.9.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.18.0
+    -   portal@3.3.10
+
+## 5.9.3
+
+### Patch Changes
+
+<sup><time>30.05.2025</time></sup>
+
+### [#1614](https://github.com/core-ds/core-components/pull/1614)
+
+-   Исправлен тип `TransitionProps` на `CSSTransitionProps` в `base-modal`
+
+## 5.9.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.17.1
+    -   portal@3.3.9
+
+## 5.9.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.17.0
+    -   portal@3.3.8
 
 ## 5.9.0